# Add bench test for mem and fix memory issue